Add constants for granularity of the parallelism levels
OpenMP-based parallelization will leverage the fact
that there are many for-loops embedded in each other, creating
a hierarchy of for-loops, wherease the iteration mostly
happens on event-sensor-hit level.
These constants will be used in the num_threads()
directives for nested OpenMP `parallel' pragmas to indicate the
parallelization dimensions. These dimensions may or may not
be adapted to the architechture the code runs on.